+++ lib/glimmer/swt/custom/drawable.rb
@@ -36,7 +36,7 @@ module Glimmer
         end
         
         def resetup_shape_paint_listeners
-          # TODO consider performance optimizations relating to order of shape rendering (affecting only further shapes not previous ones)
+          # TODO consider performance optimization relating to order of shape rendering (affecting only further shapes not previous ones)
           shapes.each do |shape|
             shape.paint_listener_proxy&.unregister
             shape.setup_paint_listener